+++ tests/ops/service/connection_config/test_mysql_connector.py
@@ -61,7 +61,7 @@ def test_mysql_connector_build_uri(connection_config_mysql, db: Session):
 def test_get_connect_args(connection_config_mysql):
     connector = MySQLConnector(configuration=connection_config_mysql)
 
-    # Default sslmode
+    # Default ssl_mode
     connection_config_mysql.secrets = {
         "username": "mysql_user",
         "password": "mysql_pw",
@@ -71,13 +71,13 @@ def test_get_connect_args(connection_config_mysql):
     connect_args = connector.get_connect_args()
     assert connect_args == {"ssl": {"mode": "preferred"}}
 
-    # Custom sslmode
+    # Custom ssl_mode
     connection_config_mysql.secrets = {
         "username": "mysql_user",
         "password": "mysql_pw",
         "host": "host.docker.internal",
         "dbname": "mysql_example",
-        "sslmode": "required",
+        "ssl_mode": "required",
     }
     connect_args = connector.get_connect_args()
     assert connect_args == {"ssl": {"mode": "required"}}